fix: move updating presets with typography to editor loader
